+++ lib/type_def/type_def.dart
@@ -1,3 +1,7 @@
+import 'package:cross_file/cross_file.dart';
+
 typedef FieldValidator<T> = String? Function(T? value);
 
 typedef TextFieldValidator = String? Function(String? value);
+
+typedef ImageFieldValidator = String? Function(XFile? file);