+++ lib/view_models/field_cubits/text_field_cubit/text_field_state.dart
@@ -33,7 +33,7 @@ class TextFieldState extends Equatable {
   }) {
     return TextFieldState(
       value: value ?? this.value,
-      errorText: errorText ?? this.errorText,
+      errorText: errorText,
       validator: validator ?? this.validator,
       isDirty: isDirty ?? this.isDirty,
     );