+++ lib/view_models/generic_cubits/field_cubit/field_cubit.dart
@@ -9,17 +9,17 @@ class FieldCubit<T> extends Cubit<FieldState<T>> {
   final FieldState<T> _initialState;
 
   FieldCubit({T? initialValue, FieldValidator<T>? validator})
-    : _initialState = FieldState<T>(
-        value: initialValue,
-        validator: validator,
-        errorText: validator?.call(initialValue),
-        isDirty: false,
-      ),
-      super(
+      : _initialState = FieldState<T>(
+    value: initialValue,
+    validator: validator,
+    errorText: null,
+    isDirty: false,
+  ),
+        super(
         FieldState<T>(
           value: initialValue,
           validator: validator,
-          errorText: validator?.call(initialValue),
+          errorText: null,
           isDirty: false,
         ),
       );
@@ -30,27 +30,26 @@ class FieldCubit<T> extends Cubit<FieldState<T>> {
       state.copyWith(
         value: value,
         errorText: error,
-        isDirty: true, // mark dirty when user updates
+        isDirty: true,
       ),
     );
   }
 
   String? validate({bool force = false}) {
-    // only validate if forced (create) or if dirty (update)
     final shouldValidate = force || state.isDirty;
     final error = shouldValidate ? state.validator?.call(state.value) : null;
+
     emit(
       state.copyWith(
-        errorText: error,
-        // once validated, we consider it dirty if forced or already dirty
-        isDirty: state.isDirty || force,
+        errorText: shouldValidate ? error : null,
+        isDirty: shouldValidate,
       ),
     );
     return error;
   }
 
   FormFieldValidator<T> get formFieldValidator =>
-      (value) => _validateAndEmit(value);
+          (value) => _validateAndEmit(value);
 
   void clear() => emit(state.init());
 
@@ -61,9 +60,10 @@ class FieldCubit<T> extends Cubit<FieldState<T>> {
     emit(
       state.copyWith(
         errorText: error,
-        isDirty: true, // validating via form field means user interacted
+        isDirty: true,
       ),
     );
     return error;
   }
 }
+